rename util.js => utils.js
