+++ lib/task.js
@@ -5,7 +5,7 @@ const streamToObservable = require('stream-to-observable');
 const Subject = require('rxjs/Subject').Subject;
 const renderers = require('./renderer');
 const state = require('./state');
-const util = require('./util');
+const utils = require('./utils');
 
 const defaultSkipFn = () => false;
 
@@ -83,7 +83,7 @@ class Task extends Subject {
 	run() {
 		const handleResult = result => {
 			// Detect the subtask
-			if (util.isListr(result)) {
+			if (utils.isListr(result)) {
 				result.setRenderer(renderers.silent);
 				this._subtasks = result.tasks;
 
@@ -100,7 +100,7 @@ class Task extends Subject {
 			}
 
 			// Detect Observable
-			if (util.isObservable(result)) {
+			if (utils.isObservable(result)) {
 				result = new Promise((resolve, reject) => {
 					result.subscribe({
 						next: data => {
