+++ test/utils.js
@@ -2,7 +2,7 @@ import test from 'ava';
 import {Observable as RxObservable} from 'rxjs/Observable';
 import ZenObservable from 'zen-observable';
 import Listr from '../';
-import {isListr, isObservable} from '../lib/util';
+import {isListr, isObservable} from '../lib/utils';
 
 test('isListr', t => {
 	t.false(isListr(null));
